Add linux-kernel-2.6.28-tt1229603
http://download.tomtom.com/open/gpl/arm11/linux-kernel-2.6.28-tt1229603.tar.gz

+++ include/sound/soc-dapm.h
@@ -304,7 +304,7 @@ struct snd_soc_dapm_widget {
 	struct list_head list;
 
 	/* dapm control */
-	short reg;						/* negative reg = no direct dapm */
+	int reg;						/* negative reg = no direct dapm */
 	unsigned char shift;			/* bits to shift */
 	unsigned int saved_value;		/* widget saved value */
 	unsigned int value;				/* widget current value */
